+++ CHANGELOG.md
@@ -12,11 +12,16 @@ This project adheres to [Semantic Versioning].
 - Variable Blake2b digest lengths and keying
 - Regular expression for testing valid hex strings
 - Node read-stream collection helper
+- Base58 validator regular expression
 
 ### Changed
 
 - Base58 helpers are namespaced
 - Concatenation helper accepts any number of inputs and can process bare numbers
+
+## Removed
+
+- The Base58 alphabet constant
 
 ## [0.1.1] - 2021-04-14